requirements-vercel.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.9.0.80
scikit-learn==1.4.0
mysql-connector-python==8.3.0
numpy==1.26.3
pandas==2.1.5
reportlab==4.1.0
vercel-wsgi==0.0.3
